Rename `app-sidebar.tsx` to `AppSidebar.tsx`
